Add approval and comments columns
This commit contains updates to the db. Two new Columns were added,
approval(boolean) and comments(text area) these are editable only by
the administrator. Therefore I made another formfile called form2,
which is for Admin only. This is accessed by accessing the request by
trying to edit it. Approval is not a checkbox.