Tweaks
The UI for the index was changed: refer to partial _blog. This was done
to make the UI nicer.
After an edit is updated, the re-route now takes the user to the root
page and posts a notice at the top.
Sort by date added to blog view, as well as Show, Edit and Destroy
Buttons.
Some css changes: seen in the section titled "Mikhail's changes"  in
the main.css file.

Precompiled before reset.